language support
By default this will check the browser's local and try to
include a js file for that locale.  English is the fallback
language.

Also, on init the page will post a message to its parent
window if one exists asking for a locale update.  The parent
window can post a message back that will update the frame's
locale.